Add a Engine method to get the 1% percentile low FPS
Measuring performance by only looking at the average FPS can be misleading,
as microstutters will not be represented in the final value.
This can lead to confusion as a game might have an average FPS of over 60,
yet it can feel very stuttery if its 1% low FPS are in the single digits.

This adds a Engine method and profiler readout for the 1% low FPS over
the last 100 rendered frames. This is a moving metric by definition,
but it still helps make these FPS drops more noticeable compared to an
average readout.