Fix Animation: reset value when adding new Bezier track
* *fixes: #81929*

In Float and Integer types, there is no subindex – only the primary value. Currently, trying to retrieve a subindex from these types in the Variant leads to a return value of null. To address this, the proposed change ensures that the default value is returned for these types instead of attempting an invalid subindex retrieval.