Add Texture Smooth Pixel Filter property to BaseMaterial3D
When enabled, this displays the texture with an appearance that mimics
nearest-neighbor filtering with a shader to smooth out aliasing between pixels.
This can be used for a pixel art look that is still antialiased
regardless of the current antialiasing method in use.

This is also available in SpriteBase3D and Label3D.